+++ lib/src/ping/base_ping.dart
@@ -25,10 +25,10 @@ abstract class BasePing {
   int? count;
 
   /// Delay between ping attempts
-  double interval;
+  int interval;
 
   /// How long to wait for a ping to return before marking it as lost
-  double timeout;
+  int timeout;
 
   // How many network hops the packet should travel before expiring
   int ttl;
